cnn-visualizer/requirements.txt
flask==3.1.3
flask-cors==6.0.2
torch==2.6.0
torchvision==0.21.0
pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless==4.13.0.92
